Fix mapping endpoint for displayContactPage method in IndexController. Adapt LoginController to be used internally with Spring Security - leave only mapping for displayLoginForm method. Move displayDashboard method to UserController. Fix login.html to have username and password fields used internally in Spring Security authentication.
