+++ src/main/java/com/mateuszmarcyk/walkthedog/user/UserController.java
@@ -1,6 +1,8 @@
 package com.mateuszmarcyk.walkthedog.user;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,4 +20,12 @@ private final UserService userService;
     public List<User> findAll() {
         return userService.getAll();
     }
+
+    @GetMapping("/dashboard")
+    public String displayDashboard(@AuthenticationPrincipal UserDetails userDetails) {
+
+        return userDetails.toString();
+    }
+
+
 }
