+++ src/main/resources/templates/login.html
@@ -20,8 +20,8 @@
       <h2 class="mb-4 text-center">Zaloguj się</h2>
       <form th:action="@{/login}" th:method="POST">
         <div class="form-group">
-          <label for="email">Adres Email</label>
-          <input type="email" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder="Wprowadź email">
+          <label for="username">Adres Email</label>
+          <input type="email" class="form-control" id="username" name="username" aria-describedby="emailHelp" placeholder="Wprowadź email">
         </div>
         <div class="form-group">
           <label for="password">Hasło</label>
